chatbot.py
When you run the program, it will prompt you to enter a statement. You can type any statement you like, and the chatbot will respond based on the keywords it finds in the statement. If it doesn't find any keywords, it will choose a response at random. If you enter "bye", the program will exit.